[DatePicker] Sort out tests
