+++ packages/material-ui-lab/src/DayPicker/DayPicker.spec.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import moment, { Moment } from 'moment';
+import DayPicker from '@material-ui/lab/DayPicker';
+
+// External components are generic as well
+<DayPicker<Moment>
+  view="date"
+  views={['date']}
+  date={moment()}
+  minDate={moment()}
+  maxDate={moment()}
+  onChange={(date) => date?.format()}
+/>;
